+++ CHANGELOG.md
@@ -49,6 +49,10 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 - Moved the role layering menu to the F1 menu (administration submenu)
   - removed the command `ttt2_edit_rolelayering`
 
+### Removed
+
+- Removed the custom loading screen (GMOD now only accepts http(s) URLs for sv_loadingurl)
+
 ### Breaking Changes
 
 - Adjusted `Player:HasRole()` and `Player:HasTeam()` to support simplified role and team checks (no parameter are supported anymore, use `Player:GetRole()` or `Player:GetTeam()` instead)